feat: Correçao de propriedades.

+++ TesteItauCorretora.Domain/Entities/Custodia.cs
@@ -8,11 +8,14 @@ namespace TesteItauCorretora.Domain.Entities;
 
 public class Custodia
 {
-    public long Id { get; set; }
-    public long ContaGraficaId { get; set; }
+    public int Id { get; set; }
+
+    public int ContaGraficaId { get; set; }
 
     public string Ticker { get; set; } = string.Empty;
+
     public int Quantidade { get; set; }
+
     public decimal PrecoMedio { get; set; }
 
     public DateTime DataUltimaAtualizacao { get; set; }
